backend/requirements.txt
asgiref==3.8.1
django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
dj-database-url==2.3.0
python-dotenv==1.1.0
yfinance==0.2.55
